More work in prog for HMM
